+++ include/gnfs/sqrt/couveignes.hpp
@@ -707,7 +707,7 @@ public:
         for (uint64_t i = 1; i < max_patterns; ++i) {
             uint64_t gray = i ^ (i >> 1);
             uint64_t changed_bit = prev_gray ^ gray;
-            size_t bit_pos = __builtin_ctzll(changed_bit);
+            size_t bit_pos = static_cast<size_t>(__builtin_ctzll(changed_bit));
             bool new_sign = (gray >> bit_pos) & 1;  // 1 = negative
 
             // Incremental CRT update + 立即归约到 [0, M-1]
@@ -974,7 +974,7 @@ public:
         for (uint64_t i = 1; i < max_patterns; ++i) {
             uint64_t gray = i ^ (i >> 1);
             uint64_t changed_bit = prev_gray ^ gray;
-            size_t bit_pos = __builtin_ctzll(changed_bit);
+            size_t bit_pos = static_cast<size_t>(__builtin_ctzll(changed_bit));
             bool new_sign = (gray >> bit_pos) & 1;
 
             for (uint32_t ci = 0; ci < d; ++ci) {
